requirements.txt
asgiref==3.11.1
Django==5.2.12
django-ckeditor-5==0.2.20
djangorestframework==3.17.1
dotenv==0.9.9
gunicorn==25.3.0
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.12.0
